refactor: extract MarkableSubmission
